feat: add tree_cell.rs
- move tree cell related code to separate file
- add handle_tree_view_event function in comms
