+++ src/comms.rs
@@ -1,10 +1,8 @@
-use std::path::Path;
-
 use gtk::glib::{self, Receiver, Sender};
-use gtk::prelude::ObjectExt;
 
 use crate::ui::notebook::editor::{get_text_buffer_by_path, set_text_on_editor};
 use crate::ui::notebook::handler::handle_notebook_event;
+use crate::ui::w_explorer::tree_view::handle_tree_view_event;
 use crate::{
     ui::{self, w_explorer::model::RootTreeModel},
     workspace::Workspace,
@@ -43,43 +41,7 @@ pub fn handle_comm_event(tx: Sender<CommEvents>, rx: Receiver<CommEvents>) {
                 handle_notebook_event(content, file_path);
             }
             CommEvents::RootTreeItemClicked(tree_model) => {
-                match tree_model {
-                    Some(tree_model) => {
-                        // Concat workspace dir path with selection
-                        let tree_item_abs_path = &tree_model.property::<String>("abs-path");
-                        let file_path = Path::new(tree_item_abs_path);
-
-                        let mut content = String::from("The selected item is not a file.");
-                        if file_path.is_file() {
-                            match std::fs::read(file_path) {
-                                Ok(data) => {
-                                    content = String::from_utf8(data)
-                                        .unwrap_or_else(|_| "File not supported".to_string());
-                                    // Update workspace's 'current open file' tracker
-                                    let open_file_path = file_path.as_os_str().to_str().unwrap();
-                                    Workspace::set_open_file_path(Some(String::from(
-                                        open_file_path,
-                                    )));
-                                }
-                                Err(error) => {
-                                    println!("Unable to read file, {}", error);
-                                }
-                            }
-                        }
-
-                        let file_path_string = String::from(file_path.to_str().unwrap());
-
-                        tx.send(CommEvents::SpawnOrFocusTab(
-                            Some(file_path_string),
-                            Some(content),
-                        ))
-                        .ok();
-                    }
-                    None => {
-                        // Reset workspace's 'current open file' tracker
-                        Workspace::set_open_file_path(None);
-                    }
-                }
+                handle_tree_view_event(tree_model, &tx);
             }
             CommEvents::UpdateRootTextViewContent(file_path, content) => {
                 G_TEXT_VIEW.with(|editor| {
