+++ src/ui/w_explorer/mod.rs
@@ -1,2 +1,3 @@
 pub mod model;
 pub mod tree_view;
+pub mod tree_cell;
